tests/unit/domain/test_stock_service.py
"""
股票服务单元测试
"""
import pytest
from unittest.mock import Mock, patch
from app.domain.services.stock_service import StockService
from app.domain.models.schemas.stock import StockResponse

class TestStockService:
    """股票服务测试类"""
    
    def setup_method(self):
        """测试前准备"""
        self.stock_service = StockService()
    
    @pytest.mark.asyncio
    async def test_get_stocks(self):
        """测试获取股票列表"""
        stocks = await self.stock_service.get_stocks(limit=5)
        assert isinstance(stocks, list)
        assert len(stocks) > 0
        assert isinstance(stocks[0], StockResponse)
    
    @pytest.mark.asyncio
    async def test_get_stock(self):
        """测试获取单个股票"""
        stock = await self.stock_service.get_stock("000001")
        # 这里根据实际实现添加断言
        pass
